libagg: use size_t for operator new argument
Fix Clang 14 build.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/5245

+++ headers/libs/agg/dbg_new/agg_dbg_new.h
@@ -26,8 +26,8 @@
 
 //#define AGG_DBG_NEW_CHECK_ADDR
 
-void* operator new (unsigned size, const char* file, int line);
-void* operator new [] (unsigned size, const char* file, int line);
+void* operator new (size_t size, const char* file, int line);
+void* operator new [] (size_t size, const char* file, int line);
 #define AGG_DBG_NEW_OPERATOR new(__FILE__, __LINE__)
 
 void  operator delete(void *ptr) throw();
